MOVEONLY: tx functions to consensus/tx_verify.o
Functions related to transaction verification.
