+++ src/Makefile.am
@@ -115,6 +115,7 @@ BITCOIN_CORE_H = \
   config.h \
   connmgr.h \
   consensus/consensus.h \
+  consensus/tx_verify.h \
   core_io.h \
   core_memusage.h \
   dosman.h \
@@ -208,6 +209,7 @@ libbitcoin_server_a_SOURCES = \
   chain.cpp \
   checkpoints.cpp \
   connmgr.cpp \
+  consensus/tx_verify.cpp \
   dosman.cpp \
   expedited.cpp \
   httprpc.cpp \
