+++ src/test/miner_tests.cpp
@@ -8,6 +8,7 @@
 #include "coins.h"
 #include "consensus/consensus.h"
 #include "consensus/merkle.h"
+#include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "main.h"
 #include "pubkey.h"
