+++ src/test/transaction_tests.cpp
@@ -8,6 +8,7 @@
 #include "test/test_bitcoin.h"
 
 #include "clientversion.h"
+#include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "core_io.h"
 #include "key.h"
