+++ src/txmempool.cpp
@@ -8,6 +8,7 @@
 
 #include "clientversion.h"
 #include "consensus/consensus.h"
+#include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "main.h"
 #include "policy/fees.h"
