+++ src/wallet/walletdb.cpp
@@ -8,6 +8,7 @@
 #include "wallet/walletdb.h"
 
 #include "base58.h"
+#include "consensus/tx_verify.h"
 #include "consensus/validation.h"
 #include "dstencode.h"
 #include "fs.h"
